notify_admin_panel/requirements.txt
python-dotenv==1.0.0
django==5.0.3
psycopg2-binary==2.9.3
uwsgi==2.0.21
django-ckeditor==6.7.1